2009-05-10	Joel Sherrill <joel.sherrill@oarcorp.com>
	* sp04/system.h, sp04/task1.c, sp04/tswitch.c, sp07/init.c,
	sp12/init.c, sp13/putbuff.c, sp13/system.h, sp13/task1.c,
	sp15/init.c, sp16/system.h, sp19/fptask.c, sp25/system.h,
	sp26/task1.c, sp27/init.c, sp28/init.c, sp29/init.c, sp31/task1.c,
	sp33/init.c, sp34/changepri.c, sp35/priinv.c, sp37/init.c,
	sp38/init.c, sp39/init.c, sp41/init.c, sp42/init.c, sp43/init.c,
	sp44/init.c, sp45/init.c, sp46/init.c, sp47/init.c, sp48/init.c,
	spfatal03/testcase.h, spfatal05/testcase.h, spfatal06/testcase.h,
	spfatal_support/system.h, spobjgetnext/init.c, spsize/getint.c,
	spsize/size.c: Fix warnings.

+++ testsuites/sptests/sp19/fptask.c
@@ -9,7 +9,7 @@
  *
  *  Output parameters:  NONE
  *
- *  COPYRIGHT (c) 1989-1999.
+ *  COPYRIGHT (c) 1989-2009.
  *  On-Line Applications Research Corporation (OAR).
  *
  *  The license and distribution terms for this file may be
@@ -52,7 +52,7 @@ rtems_task FP_task(
   printf( " - float base = (NA)\n" );
 #endif
 
-  previous_seconds = -1;
+  previous_seconds = (uint32_t)-1;
 
   while( FOREVER ) {
